+++ lib/core/posts/details_parts/src/common_post_buttons.dart
@@ -4,25 +4,20 @@ import 'package:flutter/material.dart';
 // Package imports:
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:i18n/i18n.dart';
-import 'package:oktoast/oktoast.dart';
 
 // Project imports:
-import '../../../../foundation/animations/constants.dart';
-import '../../../../foundation/clipboard.dart';
-import '../../../../foundation/toast.dart';
 import '../../../../foundation/url_launcher.dart';
 import '../../../configs/config/providers.dart';
 import '../../../configs/config/types.dart';
-import '../../../images/providers.dart';
 import '../../../premiums/providers.dart';
 import '../../../settings/routes.dart';
 import '../../../tags/show/routes.dart';
 import '../../../widgets/adaptive_button_row.dart';
-import '../../details/providers.dart';
 import '../../details_manager/routes.dart';
 import '../../post/providers.dart';
 import '../../post/routes.dart';
 import '../../post/types.dart';
+import 'toolbars/copy_post_button.dart';
 
 class CommonPostButtonsBuilder extends ConsumerWidget {
   const CommonPostButtonsBuilder({
@@ -60,10 +55,19 @@ class CommonPostButtonsBuilder extends ConsumerWidget {
     final commonButtons = [
       if (copy)
         if (config != null && configViewer != null)
-          SimpleButtonData(
-            icon: Icons.copy,
-            title: context.t.post.action.copy_image,
-            onPressed: () => copyImage(ref, post, config, configViewer),
+          ButtonData(
+            widget: CopyPostButton(
+              post: post,
+              config: config,
+              configViewer: configViewer,
+            ),
+            title: 'Copy'.hc,
+            onTap: () => showPostCopySheet(
+              context,
+              post: post,
+              config: config,
+              configViewer: configViewer,
+            ),
           ),
       if (postLinkGenerator != null && config != null)
         if (!hasStrictSFW)
@@ -113,49 +117,5 @@ class CommonPostButtonsBuilder extends ConsumerWidget {
     ];
 
     return builder(context, commonButtons);
-  }
-
-  Future<void> copyImage(
-    WidgetRef ref,
-    Post post,
-    BooruConfigAuth config,
-    BooruConfigViewer configViewer,
-  ) async {
-    void showError(String message) {
-      showErrorToast(
-        ref.context,
-        message,
-        duration: AppDurations.longToast,
-      );
-    }
-
-    final mediaUrlResolver = ref.read(mediaUrlResolverProvider(config));
-
-    final bytes = await ref.read(
-      defaultCachedImageFileProvider(
-        mediaUrlResolver.resolveMediaUrl(
-          post,
-          configViewer,
-        ),
-      ).future,
-    );
-
-    if (bytes == null) {
-      showError('Failed to get image bytes');
-      return;
-    }
-
-    try {
-      await AppClipboard.copyImageBytes(bytes);
-      showToast(
-        'Copied',
-        position: ToastPosition.bottom,
-        textPadding: const EdgeInsets.all(8),
-        duration: AppDurations.shortToast,
-      );
-    } on Exception catch (e) {
-      showError('Failed to copy image: $e');
-      return;
-    }
   }
 }